@@ -34,28 +34,37 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-		<junit-jupiter.version>5.8.2</junit-jupiter.version>
-		<junit-platform.version>1.8.2</junit-platform.version>
-		<log4j.version>2.17.1</log4j.version>
-		<mockito.version>4.3.1</mockito.version>
-		<jackson.core.version>2.13.4</jackson.core.version>
-		<jackson.databind.version>2.13.4.2</jackson.databind.version>
+		<jms-api.version>3.1.0</jms-api.version>
 
-		<spring-framework.version>5.3.18</spring-framework.version>
-		<spring-boot.version>2.5.3</spring-boot.version>
+		<spring-framework.version>6.0.3</spring-framework.version>
+
+		<log4j.version>2.19.0</log4j.version>
+
+		<jackson.core.version>2.14.1</jackson.core.version>
+		<jackson.databind.version>2.14.1</jackson.databind.version>
+
+		<qpid-jms-client.version>2.2.0</qpid-jms-client.version>
+
+		<spring-boot.version>3.0.1</spring-boot.version>
+
+		<artemis-jakarta-server.version>2.27.1</artemis-jakarta-server.version>
+
+		<junit-jupiter.version>5.8.2</junit-jupiter.version>
+		<junit-platform.version>1.9.2</junit-platform.version>
+
+		<mockito.version>4.3.1</mockito.version>
+
 		<maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
 		<maven-gpg-plugin.version>3.0.1</maven-gpg-plugin.version>
 		<maven-javadoc-plugin.version>3.2.0</maven-javadoc-plugin.version>
 		<maven-source-plugin.version>3.2.0</maven-source-plugin.version>
 		<maven-surefire-plugin.version>2.22.2</maven-surefire-plugin.version>
-
-		<jms-api.version>2.0.1</jms-api.version>
 	</properties>
 
 	<dependencies>
 		<dependency>
-			<groupId>javax.jms</groupId>
-			<artifactId>javax.jms-api</artifactId>
+			<groupId>jakarta.jms</groupId>
+			<artifactId>jakarta.jms-api</artifactId>
 			<version>${jms-api.version}</version>
 		</dependency>
 
