@@ -6,19 +6,19 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import javax.jms.Message;
+import jakarta.jms.Message;
 
 /**
- * Specifies the injection point for the value of
+ * Specifies the binding point of the value of
  * {@linkplain Message#getJMSRedelivered()} on a {@linkplain ForJmsType} object.
  * <p>
  * The injection point should be of {@code int} type.
  * <p>
  * Not applicable to {@linkplain ByJms} interfaces.
- * 
+ *
  * @author Lei Yang
  * @since 1.0
- * 
+ *
  */
 @Retention(RUNTIME)
 @Target({ ElementType.PARAMETER })
