@@ -6,10 +6,11 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import javax.jms.Message;
+import jakarta.jms.Message;
 
 /**
- * Specifies the value for {@linkplain Message#setJMSType(String)}.
+ * Specifies the binding point of the value for
+ * {@linkplain Message#setJMSType(String)}.
  * <p>
  * Can be applied to a parameter or a method on a {@linkplain ByJms} interface.
  * <p>
@@ -18,7 +19,7 @@ import javax.jms.Message;
  * When applied to a parameter, the specified value is ignored.
  * <p>
  * When applied to a method, Spring property placeholder is supported.
- * 
+ *
  * @author Lei Yang
  * @since 1.0
  */
