@@ -6,11 +6,10 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import javax.jms.Connection;
-import javax.jms.MessageProducer;
-
 import org.springframework.context.annotation.Import;
 
+import jakarta.jms.Connection;
+import jakarta.jms.MessageProducer;
 import me.ehp246.aufjms.api.dispatch.JmsDispatchFn;
 import me.ehp246.aufjms.api.jms.ConnectionFactoryProvider;
 import me.ehp246.aufjms.core.configuration.AufJmsConfiguration;
@@ -19,7 +18,11 @@ import me.ehp246.aufjms.core.dispatch.EnableByJmsBeanFactory;
 import me.ehp246.aufjms.core.dispatch.EnableByJmsRegistrar;
 
 /**
- * Enables {@link ByJms}-annotated proxy interfaces scanning.
+ * Enables client-side capabilities of Auf JMS. E.g.,
+ * {@linkplain ByJms}-annotated interfaces scanning and registration
+ * <p>
+ * By default, the package and the sub-packages of the annotated class will be
+ * scanned.
  *
  * @author Lei Yang
  * @since 1.0
@@ -64,11 +67,11 @@ public @interface EnableByJms {
      * {@linkplain Connection} retrieved from the named connection factories.
      * <p>
      * The values are passed to {@linkplain ConnectionFactoryProvider#get(String)}
-     * as connection name to create the bean.
+     * to create the beans.
      * <p>
-     * The bean name is of the form <code>JmsDistpchFn-${index}</code> where the
-     * index starts at 0. E.g., {@code JmsDispatchFn-0}, {@code JmsDispatchFn-1},
-     * and etc.
+     * The bean name is of the form <code>'jmsDispatchFn-${index}'</code> where the
+     * index starts at <code>0</code>. E.g., {@code jmsDispatchFn-0},
+     * {@code jmsDispatchFn-1}, and etc.
      * <p>
      * Does not support Spring property placeholder.
      */
