@@ -2,11 +2,14 @@ package me.ehp246.aufjms.core.dispatch;
 
 import java.util.Arrays;
 import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.config.ConstructorArgumentValues;
+import org.springframework.beans.factory.support.BeanDefinitionOverrideException;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.beans.factory.support.GenericBeanDefinition;
 import org.springframework.context.annotation.ImportBeanDefinitionRegistrar;
@@ -17,6 +20,7 @@ import me.ehp246.aufjms.api.annotation.EnableByJms;
 import me.ehp246.aufjms.api.dispatch.EnableByJmsConfig;
 import me.ehp246.aufjms.api.dispatch.JmsDispatchFn;
 import me.ehp246.aufjms.core.reflection.EnabledScanner;
+import me.ehp246.aufjms.core.util.OneUtil;
 
 public final class EnableByJmsRegistrar implements ImportBeanDefinitionRegistrar {
     private final static Logger LOGGER = LogManager.getLogger(EnableByJmsRegistrar.class);
@@ -25,29 +29,28 @@ public final class EnableByJmsRegistrar implements ImportBeanDefinitionRegistrar
     public void registerBeanDefinitions(final AnnotationMetadata metadata, final BeanDefinitionRegistry registry) {
         final var enableMap = metadata.getAnnotationAttributes(EnableByJms.class.getCanonicalName());
 
-        LOGGER.atTrace().log("Registering {} as '{}'", EnableByJmsConfig.class.getCanonicalName(),
-                EnableByJmsConfig.class.getSimpleName());
+        register(registry, beanName(EnableByJmsConfig.class), getAppConfigBeanDefinition(enableMap));
 
-        registry.registerBeanDefinition(EnableByJmsConfig.class.getSimpleName(), getAppConfigBeanDefinition(enableMap));
+        LOGGER.atTrace().log("Scanning for {}", ByJms.class::getCanonicalName);
 
-        LOGGER.atTrace().log("Scanning for {}", ByJms.class.getCanonicalName());
+        for (final var found : new EnabledScanner(EnableByJms.class, ByJms.class, metadata).perform()
+                .collect(Collectors.toList())) {
+            LOGGER.atTrace().log("Registering {}", found::getBeanClassName);
 
-        new EnabledScanner(EnableByJms.class, ByJms.class, metadata).perform().forEach(beanDefinition -> {
             final Class<?> proxyInterface;
             try {
-                proxyInterface = Class.forName(beanDefinition.getBeanClassName());
+                proxyInterface = Class.forName(found.getBeanClassName());
             } catch (final ClassNotFoundException ignored) {
                 // Class scanning started this. Should not happen.
                 throw new RuntimeException("Class scanning started this. Should not happen.");
             }
 
-            LOGGER.atTrace().log("Defining {}", beanDefinition.getBeanClassName());
+            final var beanName = Optional.ofNullable(proxyInterface.getAnnotation(ByJms.class).name())
+                    .filter(OneUtil::hasValue).orElseGet(() -> beanName(proxyInterface));
+            final var proxyBeanDefinition = this.getProxyBeanDefinition(proxyInterface);
 
-            final var name = proxyInterface.getAnnotation(ByJms.class).name();
-
-            registry.registerBeanDefinition(name.equals("") ? proxyInterface.getSimpleName() : name,
-                    this.getProxyBeanDefinition(proxyInterface));
-        });
+            register(registry, beanName, proxyBeanDefinition);
+        }
 
         final var enablerAttributes = metadata.getAnnotationAttributes(EnableByJms.class.getCanonicalName());
 
@@ -57,12 +60,24 @@ public final class EnableByJmsRegistrar implements ImportBeanDefinitionRegistrar
         }
 
         for (var i = 0; i < fns.length; i++) {
-            final var fnBeanName = JmsDispatchFn.class.getSimpleName() + "-" + i;
-
-            LOGGER.atTrace().log("Defining '{}' on '{}'", fnBeanName, fns[i]);
-
-            registry.registerBeanDefinition(fnBeanName, getFnBeanDefinition(fns[i]));
+            register(registry, "jmsDispatchFn-" + i, getFnBeanDefinition(fns[i]));
         }
+    }
+
+    private void register(final BeanDefinitionRegistry registry, final String beanName,
+            final BeanDefinition newDefinition) {
+        if (registry.containsBeanDefinition(beanName)) {
+            throw new BeanDefinitionOverrideException(beanName, newDefinition, registry.getBeanDefinition(beanName));
+        }
+
+        registry.registerBeanDefinition(beanName, newDefinition);
+    }
+
+    private String beanName(final Class<?> type) {
+        final char c[] = type.getSimpleName().toCharArray();
+        c[0] = Character.toLowerCase(c[0]);
+
+        return new String(c);
     }
 
     private BeanDefinition getAppConfigBeanDefinition(final Map<String, Object> map) {
@@ -72,14 +87,14 @@ public final class EnableByJmsRegistrar implements ImportBeanDefinitionRegistrar
         args.addGenericArgumentValue(map.get("delay"));
         args.addGenericArgumentValue(Arrays.asList((String[]) map.get("dispatchFns")));
 
-        final var proxyBeanDefinition = new GenericBeanDefinition();
-        proxyBeanDefinition.setBeanClass(EnableByJmsConfig.class);
-        proxyBeanDefinition.setConstructorArgumentValues(args);
+        final var beanDefinition = new GenericBeanDefinition();
+        beanDefinition.setBeanClass(EnableByJmsConfig.class);
+        beanDefinition.setConstructorArgumentValues(args);
 
-        proxyBeanDefinition.setFactoryBeanName(EnableByJmsBeanFactory.class.getName());
-        proxyBeanDefinition.setFactoryMethodName("enableByJmsConfig");
+        beanDefinition.setFactoryBeanName(EnableByJmsBeanFactory.class.getName());
+        beanDefinition.setFactoryMethodName("enableByJmsConfig");
 
-        return proxyBeanDefinition;
+        return beanDefinition;
     }
 
     private BeanDefinition getFnBeanDefinition(final String name) {
