@@ -4,8 +4,8 @@ import org.springframework.context.annotation.Import;
 
 import me.ehp246.aufjms.api.annotation.ByJms;
 import me.ehp246.aufjms.api.annotation.ByJms.To;
+import me.ehp246.test.EmbeddedArtemisConfig;
 import me.ehp246.aufjms.api.annotation.EnableByJms;
-import me.ehp246.aufjms.util.EmbeddedArtemisConfig;
 
 /**
  * @author Lei Yang
