@@ -5,9 +5,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.IntStream;
 
-import javax.jms.JMSException;
-import javax.jms.Session;
-
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.core.config.Configurator;
@@ -17,15 +14,17 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.support.DefaultListableBeanFactory;
 
+import jakarta.jms.JMSException;
+import jakarta.jms.Session;
 import me.ehp246.aufjms.api.endpoint.InboundEndpoint;
 import me.ehp246.aufjms.api.endpoint.InstanceScope;
 import me.ehp246.aufjms.api.endpoint.InvocableTypeDefinition;
 import me.ehp246.aufjms.api.endpoint.InvocationModel;
 import me.ehp246.aufjms.core.reflection.ReflectedType;
 import me.ehp246.aufjms.provider.jackson.JsonByJackson;
-import me.ehp246.aufjms.util.MockTextMessage;
 import me.ehp246.test.TestUtil;
 import me.ehp246.test.TimingExtension;
+import me.ehp246.test.mock.MockTextMessage;
 
 /**
  * @author Lei Yang
@@ -58,7 +57,7 @@ class DefaultInboundMessageListenerTest {
         IntStream.range(0, LOOP).forEach(i -> {
             try {
                 listener.onMessage(msg, session);
-            } catch (JMSException e) {
+            } catch (final JMSException e) {
                 throw new RuntimeException(e);
             }
         });
