@@ -2,7 +2,7 @@ package me.ehp246.aufjms.core.endpoint;
 
 import org.junit.jupiter.api.Test;
 
-import me.ehp246.aufjms.util.MockJmsMsg;
+import me.ehp246.test.mock.MockJmsMsg;
 
 /**
  * @author Lei Yang
